Initial CMake build system support
Supports out-of-tree builds without cluttering the source tree (as is
the case with autotools) and enable faster builds with Ninja.

The only missing "feature" from autotools is an uninstall target. To
mimic this after the install target: `xargs rm < install_manifest.txt`

Note that header files are not listed in the sources list, CMake
automatically discovers what headers are needed.