attrib: qualify device with an option.
I learned that the in-kernel FAT implementation in Linux supports two
ioctls, FAT_IOCTL_GET_ATTRIBUTES and FAT_IOCTL_SET_ATTRIBUTES, to get
and set file attributes on a mounted file system.

If this FUSE exFAT implementation (and/or the new in-kernel
implementation on Linux) were to implement such an ioctl in future, it
would be nice to be able to adjust this tool to support that usage
without breaking its command-line arguments in a backwards-incompatible
way. To future-proof against that, require a '-d' option to specify the
target device.

This would also allow the command to accept more that one path, like
chmod.